adjust from float to decimal in all columns to fix precision numbers problems

+++ FabricesStoreManagementSystem/DTOs/Catalog/CatalogProductRequest.cs
@@ -2,5 +2,5 @@
 
 public record CatalogProductRequest(
     Guid ProductID,
-    float Quantity
+    decimal Quantity
 );
