+++ FabricesStoreManagementSystem/Entities/StockTransaction.cs
@@ -4,7 +4,7 @@ public class StockTransaction
 {
     public Guid Id { get; set; } = Guid.CreateVersion7();
     public Guid ProductID { get; set; }
-    public float QuantityChange { get; set; }
+    public decimal QuantityChange { get; set; }
     public StockTransactionType TransactionType { get; set; }
     public Guid? ReferenceID { get; set; }
     public ReferenceTypes? ReferenceType { get; set; }
